RDKBDEV-2222: add support for GPV of Device.VoiceService.Data
Reason for change:
add support for GPV of Device.VoiceService.Data

Fix Device.X_LGI-COM_MTA. and related parameters are not retrieved in GPV of "Device."

[Description]
When bulk get is performed, if get parameter value of a parameter
fails, failure is returned from there and remaining objects are not
retrieved.

[Solution]
Add support for GPV of Device.VoiceService.Data

[Testing Performed]
GPV of  'Device.' from ACS and dmcli.

Test Procedure: Sanity.
Risks: None.

Change-Id: If54e55f4d0a70c6d117ed1249ece8cbaef0bb903
Signed-off-by: Dhanya Ravindran<dravindran.contractor@libertyglobal.com>