add support for GPV of Device.VoiceService.Data ( OFW-2106 )
Fix Device.X_LGI-COM_MTA. and related parameters are not retrieved in GPV of "Device."

[Description]
When bulk get is performed, if get parameter value of a parameter
fails, failure is returned from there and remaining objects are not
retrieved.

[Solution]
Add support for GPV of Device.VoiceService.Data

[Testing Performed]
GPV of  'Device.' from ACS and dmcli.

Signed-off-by: Dhanya Ravindran<dravindran.contractor@libertyglobal.com>